fix: allow the modal process to finish even when home changes
The bug happened because the account was selected, causing the Home page
to transfer to listing the account assets, causing the Modal component to
be garbage collected. Instead of waiting for the modal to be completed
before choosing the account (which would have moved a lot of logic
outside the modal), moving the modal ownership to the top-level app
was the least amount of change to fix the bug.

This code needs a good overhaul at some point. But this wont be the time
nor the place.

Fixes #111